Add restrictions for Prow staging process (#1844)
* add restrictions for Prow staging process

* address cr comment

* address cr issues and some fixes
