+++ images/prow-config-updater/Dockerfile
@@ -12,9 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-FROM golang:1.13
+# TODO(chizhg): Make this image smaller by only installing components we need, in this case gcloud, kubectl, bazel and Go
+FROM gcr.io/k8s-testimages/kubekins-e2e:v20200302-c921a88-master
 LABEL maintainer "Chi Zhang <chizhg@google.com>"
 RUN apt-get update
+RUN gcloud components update
+
+# Install go at 1.13 same as how it's installed in kubekins-e2e image, reference code here:
+# https://github.com/kubernetes/test-infra/blob/1e9b5dc3de4b268aab57c9c48120aa3dcf096bc6/images/kubekins-e2e/Dockerfile#L64
+ENV GO_TARBALL "go1.13.linux-amd64.tar.gz"
+RUN rm -rf /usr/local/go && \
+    wget -q "https://storage.googleapis.com/golang/${GO_TARBALL}" && \
+    tar xzf "${GO_TARBALL}" -C /usr/local && \
+    rm "${GO_TARBALL}"
 
 ENV TEMP_REPO_DIR /go/src/knative.dev/test-infra
 ENV TOOL_NAME prow-config-updater
